Add unit tests. Reorganize files. Add setup.py. Add requirements.
