+++ .gitignore
@@ -1,0 +1,8 @@
+.idea/
+dist/
+*.egg-info/
+.coverage
+*.egg
+*.pyc
+*.swp
+__pycache__
