+++ million_dollar_dream/__version__.py
@@ -1,0 +1,7 @@
+__title__ = 'million_dollar_dream'
+__description__ = 'Every man has a price.'
+__url__ = 'https://github.com/droberson/million_dollar_dream'
+__version__ = '0.0.1'
+__author__ = 'Daniel Roberson'
+__author_email__ = 'daniel(a)planethacker.net'
+__copyright__ = 'Copyright 2018 Daniel Roberson'
