+++ requirements/development.txt
@@ -1,0 +1,15 @@
+atomicwrites==1.2.1
+attrs==18.2.0
+coverage==4.5.1
+flake8==3.6.0
+mccabe==0.6.1
+more-itertools==4.3.0
+pkg-resources==0.0.0
+pluggy==0.8.0
+py==1.7.0
+pycodestyle==2.4.0
+pyfakefs==3.5
+pyflakes==2.0.0
+pytest==3.10.0
+pytest-cov==2.6.0
+six==1.11.0
