+++ requirements/production.txt
@@ -1,0 +1,1 @@
+Cython==0.29
